// src/app/model/emp/employee.ts
export class Employee {
  accountNumber: number;
  availedNoOfPaidLeaves: number = 0;
  availedNoOfSickLeaves: number = 0;
  balanceNoOfPaidLeaves: number = 0;
  balanceNoOfSickLeaves: number = 0;
  baseSalary: number = 0;
  bonus: number = 0;
  bloodGroup: string
  dateOfBirth: string;
  designation: string;
  department: string;
  dateOfJoining: string;
  email: string;
  firstName: string;
  id: number;
  lastName: string;
  phoneNumber: string;
  reportingPerson: string;
  reportingPersonId: string;
  city: string;
  country: string;
  state: string;
  streetName: string;
  streetNumber: string;
  locality: string;
  postcode: number;
  gender: string;
  emergencyContactNumber: string;
  identificationNumber: string;
  identificationType: string;
  maritialStatus: string;
  nationality: string;
  taxCountry: string;
  taxNumber: string;
  taxType: string;
  resignationDate: string;
  //salary
  houseRentAllowance: number = 0;
  performanceAllowance: number = 0;
  transportAllowance: number = 0;
  personalPay: number = 0;
  // personalTax:number = 0;
  panNumber: string;
  bankName: string;
  ifscCode: string;
  totalSalary: number = 0;
  employeeProvidentFund: number = 0;
  employerProvidentFund: number = 0;
  fuelAllowance: number = 0;
  incomeTaxAmount: number = 0;
  leaveTravelAllowance: number = 0;
  professionalTax: number = 0;
  specialAllowance: number = 0;
  performancePay: number = 0;
  noOfWorkingDays: number = 0;
  perDayRate: number = 0;
  propertyId: number;
  status: boolean;
  totalNoOfPaidLeaves: number;
  totalNoOfSickLeaves: number;
  totalNoOfUnPaidLeaves: number;
  employeeCode : string;



}